+++ packages/plugin-rax-app/src/Loaders/RouteLoader.ts
@@ -68,7 +68,6 @@ export default function (appJSON) {
   );
   const { routes, ...otherConfig } = staticConfig;
   return `
-  import { createElement } from 'rax';
   ${normalImportExpression}
   const staticConfig = ${JSON.stringify(otherConfig)};
 
@@ -121,11 +120,11 @@ function addDynamicImportRouteExpression(dynamicImports: IRoute[]): string {
     expression += `staticConfig.routes.push({
       ...${JSON.stringify(route)},
       lazy: true,
-      component: import(/* webpackChunkName: "${getComponentName(route).toLowerCase()}.chunk" */ '${getPagePathByRoute(
+      component: () => import(/* webpackChunkName: "${getComponentName(route).toLowerCase()}.chunk" */ '${getPagePathByRoute(
   route,
   { rootContext: this.rootContext },
 )}')
-      .then((mod) => mod.default || mod)
+            .then((mod) => mod.default || mod)
     });`;
   });